+++ changelog.md
@@ -7,6 +7,7 @@ Updated to MC 1.21.6
 - Dialogue screens now have a proper name, which will be narrated for players using this accessibility feature.
 This name can be changed with a new optional `name` property on dialogues. The default name in English is "Dialogue" when there is
 no interlocutor, and "Dialogue with {entity name}" when there is one.
+- The main text in dialogues now gets a scrollbar when it gets too long, and can be narrated
 
 ------------------------------------------------------
 Version 1.11.1